MA-20756 Fix the record resolution switch crash issue
on 8ulp, when display screen is portrait,
switching the resolution of record, then back to preview,
the layout of the BottomBar will be re-created,
there is a low probability of getting '0' value of
measured width and height, in this case, values will not be saved,
which causes the exception in class View.

Add 'super.onMeasure' under this condition inside the
override onMeasure to save the measured width and height.
('super.onMeasure' will call the 'setMeasuredDimension',
this method must be called by onMeasure(int, int) to
store the measured width and measured height.)

Change-Id: I56f02ada30230ebfabad0999e8299c79223f6af4
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/20304
Reviewed-by: fang hui <hui.fang@nxp.com>