MA-21080-2 fix monkey crash 'divide by zero' in Camera2
This function appears to map a value within
arange (eg, 0.5 in [0, 1]) to scale to some
output range (eg, 1.5 in [1, 2] or 2 in [1, 3])),
the zero appears when in_min = in_max, and make no sense here.
Ensures the return value fits in a
given numerical range [out_min, out_max] is more reasonable.

Change-Id: I54c7cfb9d82f0a336adf5d3a37d4827f2688e94c
Signed-off-by: Anle Pan <anle.pan@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/22270
Reviewed-by: Hui Fang <hui.fang@nxp.com>