MA-21638 Fix Camera Intents test fails in step3
The issue only appears on 8ulp with panel, which
was portrait screen and with 720p resolution.
Camera Intents test in step3 will launch
camera app to take photo, after the capturing,
the done_button does not appear, making it
impossible to confirm and conduct the subsequent testing.

To fix the issue, add layout_weight to both
retake_button and cancel_button to ensure that
all three buttons can display normally.

Change-Id: I807d72c827ae99b41684dac2762f7113d20d8b08
Signed-off-by: Anle Pan <anle.pan@nxp.com>